오늘자 전체 병합 완료

+++ src/main/java/com/example/CafeAutoSystem/vendoringredient/service/VendorIngredientService.java
@@ -30,7 +30,7 @@ public class VendorIngredientService {
                 .vendorIngredientId(vi.getVendorIngredientId())
                 .vendorId(vi.getVendor() != null ? vi.getVendor().getVendorId() : null)
                 .vendorName(vi.getVendor() != null ? vi.getVendor().getVendorName() : null)
-                .ingredientId(vi.getIngredient() != null ? vi.getIngredient().getIngredientId() : null)
+                .ingredientId(vi.getIngredient() != null ? vi.getIngredient().getIngredientId().intValue() : null)
                 .ingredientName(vi.getIngredient() != null ? vi.getIngredient().getIngredientName() : null)
                 .unitPrice(vi.getUnitPrice())
                 .priorityRank(vi.getPriorityRank())
